refactor tooltip
move edit functionality to theme
