+++ themes/base.js
@@ -5,6 +5,7 @@ import Theme from '../core/theme';
 import ColorPicker from '../ui/color-picker';
 import IconPicker from '../ui/icon-picker';
 import Picker from '../ui/picker';
+import Tooltip from '../ui/tooltip';
 import icons from '../ui/icons';
 
 
@@ -165,6 +166,77 @@ BaseTheme.DEFAULTS = {
 };
 
 
+class BaseTooltip extends Tooltip {
+  constructor(quill, boundsContainer) {
+    super(quill, boundsContainer);
+    this.textbox = this.root.querySelector('input[type="text"]');
+    this.listen();
+  }
+
+  listen() {
+    this.textbox.addEventListener('keydown', (event) => {
+      if (Keyboard.match(event, 'enter')) {
+        this.save();
+        event.preventDefault();
+      } else if (Keyboard.match(event, 'escape')) {
+        this.cancel();
+        event.preventDefault();
+      }
+    });
+  }
+
+  cancel() {
+    this.hide();
+  }
+
+  edit(mode = 'link', preview = null) {
+    this.root.classList.remove('ql-hidden');
+    this.root.classList.add('ql-editing');
+    if (preview != null) {
+      this.textbox.value = preview;
+    } else if (mode !== this.root.dataset.mode) {
+      this.textbox.value = '';
+    }
+    this.textbox.select();
+    this.textbox.setAttribute('placeholder', this.textbox.dataset[mode] || '');
+    this.root.dataset.mode = mode;
+    this.position(this.quill.getBounds(this.quill.selection.savedRange));
+  }
+
+  save() {
+    switch(this.root.dataset.mode) {
+      case 'link':
+        let url = this.textbox.value;
+        let scrollTop = this.quill.root.scrollTop;
+        if (this.linkRange) {
+          this.quill.formatText(this.linkRange, 'link', url, Emitter.sources.USER);
+          delete this.linkRange;
+        } else {
+          this.quill.focus();
+          this.quill.format('link', url, Emitter.sources.USER);
+        }
+        this.quill.root.scrollTop = scrollTop;
+        break;
+      case 'formula':  // fallthrough
+      case 'video':
+        let range = this.quill.getSelection(true);
+        let index = range.index + range.length;
+        if (range != null) {
+          this.quill.insertEmbed(index, this.root.dataset.mode, this.textbox.value, Emitter.sources.USER);
+          if (this.root.dataset.mode === 'formula') {
+            this.quill.insertText(index + 1, ' ', Emitter.sources.USER);
+          }
+          this.quill.setSelection(index + 2, Emitter.sources.USER);
+        }
+        break;
+      default:
+    }
+    this.textbox.value = '';
+    this.hide();
+  }
+}
+
+
 function fillSelect(select, values, defaultValue = false) {
   values.forEach(function(value) {
     let option = document.createElement('option');
@@ -178,4 +250,4 @@ function fillSelect(select, values, defaultValue = false) {
 }
 
 
-export default BaseTheme;
+export { BaseTooltip, BaseTheme as default };
